requirements.txt
# Mac M1 / 3090 通用依赖
pandas>=2.0
numpy>=1.24,<2.0
openpyxl>=3.1
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
matplotlib-venn>=0.11
wordcloud>=1.9
tqdm>=4.65
fasttext-wheel>=0.9.2

# 深度学习相关（Mac M1 可装，CPU/MPS 推理；3090 装 GPU 版 torch 即可）
torch>=2.1
transformers>=4.40
accelerate>=0.27
sentencepiece
protobuf
modelscope>=1.13
